+++ src/test/kotlin/cz/startnet/utils/pgdiff/PgDiffTest.kt
@@ -1,6 +1,10 @@
 package cz.startnet.utils.pgdiff
 
+import io.kotest.matchers.collections.shouldBeEmpty
+import io.kotest.matchers.shouldBe
+import io.kotest.matchers.string.shouldBeBlank
 import io.kotest.matchers.string.shouldBeEmpty
+import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ArgumentsSource
 
@@ -16,7 +20,8 @@ class PgDiffTest {
     fun runDiffSameOriginal(testFiles: SQLDiffTestFiles) {
         val dump = testFiles.old.readText()
         val diff = PgDiff.createDiff(dump, dump)
-        diff.trim { it <= ' ' }.shouldBeEmpty()
+        diff.script.shouldBeBlank()
+        diff.diffIgnored().shouldBeEmpty()
     }
 
     /**
@@ -29,7 +34,8 @@ class PgDiffTest {
     fun runDiffSameNew(testFiles: SQLDiffTestFiles) {
         val dump = testFiles.new.readText()
         val diff = PgDiff.createDiff(dump, dump)
-        diff.trim { it <= ' ' }.shouldBeEmpty()
+        diff.script.shouldBeBlank()
+        diff.diffIgnored().shouldBeEmpty()
     }
 
     /**
@@ -42,6 +48,22 @@ class PgDiffTest {
     fun runDiff(testFiles: SQLDiffTestFiles) {
         val old = testFiles.old.readText()
         val new = testFiles.new.readText()
-        PgDiff.createDiff(old, new)
+        val diff = PgDiff.createDiff(old, new)
+        diff.diffIgnored().joinToString("\n").shouldBeEmpty()
+    }
+
+    @Test
+    fun testIgnoredDiffer() {
+        val old = "CREATE SERVER myserver FOREIGN DATA WRAPPER postgres_fdw OPTIONS (dbname 'foodb');"
+        val new = "CREATE SERVER myserver FOREIGN DATA WRAPPER postgres_fdw OPTIONS (dbname 'bardb');"
+        val diff = PgDiff.createDiff(old, new)
+        diff.script.shouldBeBlank()
+        diff.diffIgnored().joinToString("\n") shouldBe """
+            --- old
+            +++ new
+            @@ -1,1 +1,1 @@
+            -CREATE SERVER myserver FOREIGN DATA WRAPPER postgres_fdw OPTIONS (dbname 'foodb');
+            +CREATE SERVER myserver FOREIGN DATA WRAPPER postgres_fdw OPTIONS (dbname 'bardb');
+            """.trimIndent()
     }
 }