+++ src/test/resources/pgdiff_test_files/add_cluster_original.sql
@@ -1,28 +1,3 @@
---
--- PostgreSQL database dump
---
-
-SET client_encoding = 'UTF8';
-SET check_function_bodies = false;
-SET client_min_messages = warning;
-
---
--- Name: SCHEMA public; Type: COMMENT; Schema: -; Owner: postgres
---
-
-COMMENT ON SCHEMA public IS 'Standard public schema';
-
-
-SET search_path = public, pg_catalog;
-
-SET default_tablespace = '';
-
-SET default_with_oids = false;
-
---
--- Name: testtable; Type: TABLE; Schema: public; Owner: postgres; Tablespace:
---
-
 CREATE TABLE testtable (
     field1 integer,
     field2 integer,