+++ src/test/resources/pgdiff_test_files/add_column_add_defaults_original.sql
@@ -39,25 +39,3 @@ CREATE TABLE table1 (
 
 ALTER TABLE public.table1 OWNER TO postgres;
 
---
--- Data for Name: table1; Type: TABLE DATA; Schema: public; Owner: postgres
---
-
-COPY table1 (col1) FROM stdin;
-\.
-
-
---
--- Name: public; Type: ACL; Schema: -; Owner: postgres
---
-
-REVOKE ALL ON SCHEMA public FROM PUBLIC;
-REVOKE ALL ON SCHEMA public FROM postgres;
-GRANT ALL ON SCHEMA public TO postgres;
-GRANT ALL ON SCHEMA public TO PUBLIC;
-
-
---
--- PostgreSQL database dump complete
---
-