+++ src/test/resources/pgdiff_test_files/add_owned_sequence_original.sql
@@ -74,25 +74,3 @@ SELECT pg_catalog.setval('table1_col1_seq', 1, false);
 ALTER TABLE ONLY table1 ALTER COLUMN col1 SET DEFAULT nextval('table1_col1_seq'::regclass);
 
 
---
--- Data for Name: table1; Type: TABLE DATA; Schema: public; Owner: postgres
---
-
-COPY table1 (col1) FROM stdin;
-\.
-
-
---
--- Name: public; Type: ACL; Schema: -; Owner: postgres
---
-
-REVOKE ALL ON SCHEMA public FROM PUBLIC;
-REVOKE ALL ON SCHEMA public FROM postgres;
-GRANT ALL ON SCHEMA public TO postgres;
-GRANT ALL ON SCHEMA public TO PUBLIC;
-
-
---
--- PostgreSQL database dump complete
---
-