+++ src/test/resources/pgdiff_test_files/add_type_new.sql
@@ -20,10 +20,6 @@ CREATE TYPE descr_type AS (
 
 ALTER TYPE descr_type OWNER TO postgres;
 
-SET default_tablespace = '';
-
-SET default_with_oids = false;
-
 --
 -- TOC entry 182 (class 1259 OID 16390)
 -- Name: t1; Type: TABLE; Schema: public; Owner: dv